Decouples PhysicalTable name from Druid table name
--Previously, the name of a `PhysicalTable` was one and the same as the
name of the associated datasource in Druid. However, this inhibits the
ability of Fili users to make creative use of Physical Tables to model
unusual Druid configurations.

--Therefore, the Fili name of a `PhysicalTable` has been decoupled from
the Druid name. The Fili name (`name`) is used to uniquely identify the
table in Fili, but the Druid name is the name serialized when building a
druid query. It is also the name used when querying Druid for the segment
metadata of a table.