Removes `RequestLog::switchTiming`.
--`RequestLog::switchTiming` works by stopping the most recent timer,
and then starting the specified timer. As a result, we can fail to stop
a timer A if we start and stop timer B inside a method invoked while
timer A is running. As a result, in order to ensure that all of your
timers are working correctly, you need to be cognizant of the internal
workings of every method you invoke within the timed block.

--Therefore, in an effort to make timed code more modular, we've removed
the `RequestLog::switchTiming` method, and replaced all instances of it
with try-finally blocks that explicitly stop and start each timer. It's
ugly, but it's also much less brittle.