Expose the durations stored in the RequestLog.
--Some customers may have their own custom logging format that differs
significantly from how the `BardLoggingFilter` logs the `RequestLog`.
However, they may still want access to the many timings that the
`RequestLog` tracks.

The ideal solution is to rearchitect the
`RequestLog` to make it easy to replace with a custom logging scheme
and/or make it truly a-la carte. However, as a stopgap, simply
providing a means of getting at the durations is sufficient.