Stops using `IndexSearcher::count` in Lucene.
--It looks like `IndexSearcher::count` fires off an entire query, and
then counts the number of documents that were hit. Currently, we are
using it to determine the total number of dimension rows that satisfy a
given query. However, this is unnecessary and wasteful, because the
`TopDocs` object returned by the query for the actual data _also_
contains information on the total number of documents hit by the query.