Improves `Dimension::getFieldByName` documentation
-- The `ApiFilter` constructor relies on the fact that
`Dimension::getFieldByName` throws an `IllegalArgumentException` to
perform validation. However, that is not documented on the `Dimension`
interface, which may lead implementors of `Dimension` to erroneously
assume that returning `null` is valid if a `DimensionField` is not found
(which is common for this kind of method). However, what actually
happens is that the `ApiFilter` is constructed with a `null` field,
leading to a `NullPointerException` at some point later.`

-- So the documentation on `Dimension` is updated to make that explicit.