Add injection point for handling exceptions (#742)
* Add injection point for handling exceptions

-- Before we hard-coded how we handled exceptions in the Servlets.
For example, by default we return a 400 - Bad Request if we get an
Exception. However, this is very inflexible. For example, suppose
customers have some customizations (like their own SearchProviders) that
may generate errors the DataServlet doesn't know to look for, but aren't
the fault of the user. Then, Fili currently returns a 400, even though
it should probably be some sort of 5xx.

So we add an injection point to all servlets that allows customers to
provide their own logic for handling DataServlet errors.

-- Tests have been added for unit tests, but not the direct test on
`DataServlet` that Michael recommended because of how brittle such a
test would be.