First pass at implementing Luthier.
-- Currently, Fili relies on Java-based configuration. Each metric,
    dimension, and table is individually defined and built in Java.
    Unfortunately, the way we have configuration currently set up, there
    are several pain points:

    1. It's very hard to generate metrics and dimensions based on a
    common structure. So if you need to generate metrics based on other
    metrics (say you want a dailyAvg metric for a bunch of your other
    metrics), you end up with a *huge* number of very similar metrics
    taht each must be individually configured, maintained and
    understoo.d

    2. Even very simple configuration is buried under Java verbosity,
    making the configuration difficult to read, update, and understand.

Therefore, we need a much simpler, concise and clearer language for
doing configuration. However, it can't just be a static data format like
JSON or YAML, because we need to support being able to programatically
generate metrics to satisfy usecase 1.

Thererfore, we decided to use Lua. Lua was in fact originally developed
to handle a very similar usecase. It was built to be a data
specification language with support for simple transformations and
generation, which is then loaded into a "heavier duty" program.

The language itself is *extremely* small and easy to learn.
Configuration is specified in a table (dictionaries), which is Lua's
only complex data structure.

While we may someday support calling the Lua code directly from Fili, as
of now, the Lua generates a JSON file that is then read by Fili using
Jackson. Based on the resulting JSON, Fili automatically constructs and
registers the tables, metrics, and dimensions.

However, there is sometimes a *little* bit of Java code that needs to be
written. If a customer has a custom search provider, key value store, or
metric maker, then they need to register a provider for those objects with the
appropriate dictionary in Fili. Then, they can reference the search
provider, key value store or metric maker by name in their Lua code.