Adds a custom ctags file for Lua.
-- (Exuberant) Ctags is a program that lets you look up the definition
of a symbol: http://ctags.sourceforge.net/. It comes with builtin
support for Lua, but its tag generation isn't that great (for one thing,
it includes the module name in a function's tag). This renders the tag all but
useless, especially if you follow the standard convention of `M` for the
current file's module. The included .ctags file allows you to generate
tags that contain *just* the function's name. While this means you lose
any namespacing, it does mean you can easily look up all possible
definitions of a symbol. In a program as small as Fili configuration,
the lack of namespacing shouldn't be a big problem.