Removes static on a bunch of helper methods.
-- The ConjunctionDruidFilterBuilder has a whole bunch of protected
helper methods that are meant to be used only by it and its children.
However, those methods are all static, meaning they can't be overridden
by customers. This makes the ConjunctionDruidFilterBuilder *very* rigid,
and removes a lot of the benefit of having protected helper methods for
each step in the first place.

-- This PR makes those helper methods instance methods.

-- It also corrects a typo in the DruidInFilterBuilder documentation.