Use DateTime for expiration.
-- Before, we were using a vague "seconds" but what that "seconds" meant
varied, partially because memcache gave you two options: Unix time, or
relative time up to 30 days from now. This is a poor choice for an
interface that could be put in front of any sort of cache. Instead, we
just make it a DateTime: the date on which the cache should entry
should expire. Clear and unambiguous.