Adds key-specific expiration to DataCache. (#911)
* Adds key-specific expiration to DataCache.

-- We have a few customers who would like to be able to set different
expiration times for different values in Fili's response cache. This
PR adds a new method to the DataCache interface
`boolean set(String key, T value, int expiration)` that allows people
to set the expiration for the specified key at cache insertion time.
The default implementation drops the expiration, so customers who have
their own DataCache implementation will be unaffected.

-- We also implmenet the method in `MemDataCache` and delegate the
old `set(String key, T value)` to it, using the configured
`EXPIRATION` as the expiration time.

* Use DateTime for expiration.

-- Before, we were using a vague "seconds" but what that "seconds" meant
varied, partially because memcache gave you two options: Unix time, or
relative time up to 30 days from now. This is a poor choice for an
interface that could be put in front of any sort of cache. Instead, we
just make it a DateTime: the date on which the cache should entry
should expire. Clear and unambiguous.