A few more tweaks.
-- Changes (hopefully improving) some JavaDoc, including fixing their
styling:
        a. When wrapping a `@param` description, the second line starts
            right underneath `@param`.
        b. There should be two spaces between the name of the param, and
            its description.

-- Includes the list of fields we know about for a dimension in the
`unknownDefaultField` error message. Keep in mind that customers may be
generating their dimension fields (or anything really) programatically,
so if we can include our view of the world (well, the pertinent bit) in
the error message, it'll go a long way to helping customers fix their
config.

-- Adds validation for every dimension field. Before, if a customer
forgot a field that we relied on, they'd get a *very* not helpful
`NullPointerException`. Now, we tell them which field is missing in
which dimension. Calling this method is before every field request in
every builtin factory will be less than ideal, and will become
unnecessary if/when we do the bean validation thing that @mclawhor
recommended, but that isn't here yet, and the config should be useable
even if we don't have time to do add bean validation anytime soon
(which is likely).