Makes NoMatchFoundException a RuntimeException.
-- Currently, NoMatchFoundException is a checked Exception. At one
point, this may have been what we wanted. However, ever since we
introduced the `DataExceptionHandler` class, Fili already
handles the NoMatchFoundException by default at the top level, so
there's no need for callers to explicitly handle it. They can just let
it bubble up the callstack.

-- As of right now, it doesn't matter because the NoMatchFoundException
is only thrown in the DataServlet. However, one customer is trying to
build a Druid Query in a RequestHandler. While building a Druid Query,
it's possible that `NoMatchFoundException` will be thrown. However, we
can't just it bubble up because it's a checked exception, and
`RequestHandler.handle` doesn't throw `NoMatchFoundException`. We
can't wrap it in a RuntimeException either, because in that case
the FiliDataExceptionHandler won't recognize it as a
`NoMatchFoundException` and therefore won't print a user-friendly error
message.

-- Therefore, we make `NoMatchFoundException` a RuntimeException.
Not only don't we *have* to handle `NoMatchFoundException` at
the place where it's thrown (since by default it's handled higher up
the callstack), but it also inhibits composability.