Adds a protected method to DruidQueryBuilder to determine query type.
-- There may be situations where customers of Fili need to tweak the
rules on when Fili does or doesn't optimize a given DataApiRequest into
a topN, timeseries or other query type. However, currently the
optimization code is fairly interwoven, and not particularly isolated.

-- This PR isolates the key logic for determining the query type to
build, and pulls it out into a separate protected method that returns
the query type that Fili should build. Now, if customers need to modify
the rules for what query type to build, they can override the given
protected method.