Adds a method to DataApiRequest to turn query optimization on/off.
-- Fili currently will attempt to use the most efficient Druid query
type that can answer the given query. While this is wonderful, it
doesn't play well with attempts to support sending multiple Druid
queries from a single ApiRequest, because it becomes theoretically
possible for a single ApiRequest to create multiple different types of
queries, and keeping track of all those types so that their responses
can be parsed correctly is painful.

-- Therefore, we include a new method on DataApiRequest that indicates
whether or not this request should be optimized, or should generate a
GroupBy query, even if it could be satisfied by a timeseries or topN.
This allows customers to force their query to follow a particular
pattern if they so desire.

-- We also modified DataApiRequestImpl to include a boolean switch that
can be flipped on and off. This was not added to the interface in order
to avoid disrupting customers who aren't concerned with such
functionality.