Unifies DataApiRequest and TablesApiRequest under ApiRequestWithTable.
-- It is not unusual for customers to apply security filters based on
the table in the request. Sometimes, these rules (especially for
rules that control whether customers can even *see* a given table) need
to be applied to both the data and tables endpoint.

However, as things are currently structured, this necessitates code
duplication, or handrolled checks for the tables endpoint because the
RoleBasedTableValidatorRequestMapper only applies to DataApiRequests.
So a new interface `ApiRequestWithTable` is introduced that extends
`ApiRequest` and adds the `getTable` method. This allows us to
generalize the `RoleBasedTableValidatorRequestMapper` (and possibly
others if necessary) to both DataApiRequests and TablesApiRequests.