Reworks role based resource validation.
-- We fully generalize RoleBasedTableValidatorRequestMapper to apply to
any ApiRequest, not just DataApiRequest, and in the process, rename it
to `RoleBasedValidatorRequestMapper` since it no longer
necessarily involves tables.

-- RoleBasedTableValidatorRequestMapper becomes a subclass of
`RoleBasedValidatorRequestMapper` specifically for TableApiRequests.
The subclass delegates to its parent if the request contains a table
(i.e. the user requests information about a specific table). Otherwise,
if there is no table on the request (i.e. the user asks for
information about all tables), the
`RoleBasedTableValidatorRequestMapper` instead modifies the
TableApiRequest to only expose those tables that pass validation.